feat: use groups in storage and create error handling

+++ src/storage/groups/groupCreate.ts
@@ -2,16 +2,17 @@ import AsyncStorage from '@react-native-async-storage/async-storage'
 
 import { StorageKey } from '@storage/storageConfig'
 import { groupGetAll } from './groupGetAll'
+import { AppError } from '@utils/AppError'
 
 export async function groupCreate(newGroup: string) {
-  try {
-    const storedGroups = await groupGetAll()
+  const storedGroups = await groupGetAll()
 
-    const storage = JSON.stringify([...storedGroups, newGroup])
-    await AsyncStorage.setItem(StorageKey.GROUP_COLLECTION, storage)
-  } catch (error) {
-    console.error(error)
+  const groupAlreadyExists = storedGroups.includes(newGroup)
 
-    throw error
+  if (groupAlreadyExists) {
+    throw new AppError('Já existe uma turma cadastrada com esse nome.')
   }
+
+  const storage = JSON.stringify([...storedGroups, newGroup])
+  await AsyncStorage.setItem(StorageKey.GROUP_COLLECTION, storage)
 }
